+++ script/socket_repl.py
@@ -1,14 +1,21 @@
 #! /usr/bin/env python3
-import os, subprocess
+import os, random, subprocess
 
 if __name__ == '__main__':
-  print("Starting Server Socket REPL at port 5555")
   os.chdir(os.path.dirname(__file__) + "/..")
+  
+  port = random.randint(1025, 65535)
+  print(f"Starting Server Socket REPL at port {port}", flush = True)
+  with open(".repl-port", "w") as file:
+    file.write(str(port))
+  
   subprocess.check_call(['clojure',
     '-J--add-opens=java.base/java.io=ALL-UNNAMED',
     '-X', 'clojure.core.server/start-server',
     ':name', 'repl',
-    ':port', '5555',
+    ':port', str(port),
     ':accept', 'clojure.core.server/repl',
     ':server-daemon', 'false'
   ])
+
+  os.remove(".repl-port")