+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### WIP
 
+- New feature: Watches!
+- Added `Add Watch` command
 - Fixed `Reconnect` command
 - Added `output.repl` panel for raw nREPL output #104
 - Added `Toggle Output Panel` command for raw nREPL connections #104